If properties are ignored , do not try to add property
As per the docs:
"When using property to declare a dependency, do not specify a type or a default: DI/1 assumes that typed properties (and defaulted properties) are intended to generate specific getters and setters on transients or for ORM integration, rather than just dependencies. "

In the metadata, the property can be ignored if it has a type or default. This is to check if it is ignored and do nothing, regardless of